requirements.txt
aiogram
aiohttp
aiosqlite
python-dotenv
APScheduler
